Added new way to sign urls through impersonation API
Thanks to https://github.com/salrashid123/gcpsamples/pull/17#issuecomment-894224128
