+++ gcs_keyless_signedurl/requirements.txt
@@ -2,3 +2,4 @@ requests
 google-api-python-client
 httplib2
 oauth2client
+google-cloud-storage
